Update CI/CD implementation

+++ scripts/deploy-image.sh
@@ -2,15 +2,8 @@
 set -e
 echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin
 
-mv services ../
-mkdir services
-mv ../services/$1 services/
-
-cp "services/$1/Dockerfile" Dockerfile
-docker build -q -t "$1:$2" --build-arg SERVICE=$1 .
-
-mv ../services/* services/
-rm -rf ../services
+[[ -f "services/$1/Dockerfile" ]] && cp "services/$1/Dockerfile" Dockerfile
+docker build -t "$1:$2" --build-arg SERVICE=$1 .
 
 docker tag "$1:$2" "basecms/identity-x-$1-service:$2"
 docker push "basecms/identity-x-$1-service:$2"
